+++ FireworksMania/Scripts/Core/Common/PlayerSpawnLocation.cs
@@ -2,8 +2,15 @@
 
 namespace FireworksMania.Core.Common
 {
-    [AddComponentMenu("Fireworks Mania/Common/PlayerSpawnLocation")]
     public class PlayerSpawnLocation : MonoBehaviour
     {
+
+#if UNITY_EDITOR
+        private void OnDrawGizmos()
+        {
+            FireworksMania.Core.Utilities.GizmosUtility.DrawArrow(this.transform.position, this.transform.forward, Color.gray, 1f, 0.25f);
+        }
+#endif
+
     }
 }